move symbols declaration to their own file, might end up be generated automatically later somehow.
